requirements.txt
djangorestframework>=3.0.0
django>=1.8
